Move profile classes to deob-util
This allows deob-ast to pull information from the profile.
