+++ deob-bytecode/src/main/kotlin/org/openrs2/deob/bytecode/transform/VisibilityTransformer.kt
@@ -18,8 +18,8 @@ import org.openrs2.asm.classpath.Library
 import org.openrs2.asm.filter.MemberFilter
 import org.openrs2.asm.filter.UnionMemberFilter
 import org.openrs2.asm.transform.Transformer
-import org.openrs2.deob.bytecode.Profile
 import org.openrs2.deob.bytecode.filter.ReflectedConstructorFilter
+import org.openrs2.deob.util.profile.Profile
 import org.openrs2.util.collect.DisjointSet
 
 @Singleton
